refactor: UserAgentTableをコンポーネント化
