+++ frontend/src/features/userAgent/index.ts
@@ -1,4 +1,5 @@
 export * from './components/CreateUserAgentButton'
 export * from './components/UpdateUserAgentButton'
+export * from './components/UserAgentTable'
 export * from './useUserAgent'
 export * from './types'
